implement RoutePatternSelect using downshift
