+++ app/component/routepage/RoutePatternSelect.js
@@ -1,26 +1,13 @@
-/* eslint-disable no-param-reassign */
-/* eslint-disable func-names */
-/* eslint-disable import/no-unresolved */
+import React from 'react';
 import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import {
-  createRefetchContainer,
-  fetchQuery,
-  graphql,
-  ReactRelayContext,
-} from 'react-relay';
+import { useSelect } from 'downshift';
+import { Link } from 'found';
+import { FormattedMessage } from 'react-intl';
 import cx from 'classnames';
-import sortBy from 'lodash/sortBy';
-import { routerShape, Link } from 'found';
-import Autosuggest from 'react-autosuggest';
-import connectToStores from 'fluxible-addons-react/connectToStores';
-import { enrichPatterns } from '@digitransit-util/digitransit-util';
-import { FormattedMessage, intlShape } from 'react-intl';
-import { routeShape, relayShape, configShape } from '../../util/shapes';
 import Icon from '../Icon';
-import { routePagePath, PREFIX_STOPS } from '../../util/path';
+import { routePagePath } from '../../util/path';
 import { addAnalyticsEvent } from '../../util/analyticsUtils';
-import { unixToYYYYMMDD } from '../../util/timeUtils';
+import { configShape, patternShape } from '../../util/shapes';
 
 function patternOptionText(pattern) {
   if (pattern) {
@@ -34,7 +21,7 @@ function patternOptionText(pattern) {
   return '';
 }
 
-function patternTextWithIcon(pattern) {
+export function patternTextWithIcon(pattern) {
   if (pattern) {
     const text = patternOptionText(pattern);
     const i = text.search(/➔/);
@@ -53,416 +40,207 @@ function patternTextWithIcon(pattern) {
   return null;
 }
 
-function filterSimilarRoutes(routes, currentRoute) {
-  const withoutCurrent = routes.filter(r => r.gtfsId !== currentRoute.gtfsId);
-
-  let routeBasename = currentRoute.shortName;
-  if (Number.isNaN(Number(routeBasename))) {
-    routeBasename = routeBasename.replace(/\D/g, ''); // Delete all non-digits from the string
-  }
-  const onlyRelatedRoutes = withoutCurrent.filter(r =>
-    Number.isNaN(Number(r.shortName.replace(routeBasename, '')[0])),
-  );
-  return sortBy(onlyRelatedRoutes, 'shortName');
-}
-
-function renderPatternSelectSuggestion(item, currentPattern) {
-  if (item.stops) {
+const PatternOption = ({
+  option,
+  optionIndexTable,
+  highlightedIndex,
+  getItemProps,
+  currentPattern,
+}) => {
+  if (option.stops) {
     return (
-      <>
-        {patternTextWithIcon(item)}
-        {item.code === currentPattern.code && (
-          <>
-            <Icon className="check" img="icon_check" />
-            <span className="sr-only">
-              <FormattedMessage id="route-page.pattern-chosen" />
-            </span>
-          </>
+      <li
+        aria-label={patternOptionText(option)}
+        key={option.code}
+        {...getItemProps({
+          item: option,
+          className: cx(
+            'suggestion',
+            optionIndexTable[option.code] === highlightedIndex &&
+              'suggestion--highlighted',
+          ),
+        })}
+      >
+        {patternTextWithIcon(option)}
+        {option.code === currentPattern.code && (
+          <Icon aria-hidden="true" className="check" img="icon_check" />
         )}
-      </>
+      </li>
     );
   }
-  if (item.shortName && item.longName && item.mode) {
-    const lowerCaseItemMode = item.mode.toLowerCase();
+  if (option.shortName && option.longName && option.mode) {
     return (
-      <Link
-        to={routePagePath(item.gtfsId)}
-        onClick={e => {
-          e.stopPropagation();
-        }}
+      <li
+        {...getItemProps({
+          item: option,
+          className: cx(
+            'suggestion',
+            optionIndexTable[option.gtfsId] === highlightedIndex &&
+              'suggestion--highlighted',
+          ),
+        })}
       >
-        <div className="similar-route">
-          <Icon
-            className={lowerCaseItemMode}
-            img={`icon_${lowerCaseItemMode}`}
-            color={item.color ? `#${item.color}` : null}
-          />
-          <div className="similar-route-text">
-            <span className="similar-route-name">{item.shortName}</span>
-            <span className="similar-route-longname">{item.longName}</span>
-          </div>
-          <div className="similar-route-arrow-container">
+        <Link
+          to={routePagePath(option.gtfsId)}
+          onClick={e => {
+            e.stopPropagation();
+          }}
+        >
+          <div className="similar-route">
             <Icon
-              className="similar-route-arrow"
-              img="icon_arrow-collapse--right"
+              className={option.mode.toLowerCase()}
+              img={`icon_${option.mode.toLowerCase()}`}
+              color={option.color ? `#${option.color}` : null}
             />
+            <div className="similar-route-text">
+              <span className="similar-route-name">{option.shortName}</span>
+              <span className="similar-route-longname">{option.longName}</span>
+            </div>
+            <div className="similar-route-arrow-container">
+              <Icon
+                className="similar-route-arrow"
+                img="icon_arrow-collapse--right"
+              />
+            </div>
           </div>
-        </div>
-      </Link>
+        </Link>
+      </li>
     );
   }
   return null;
-}
-
-class RoutePatternSelect extends Component {
-  constructor(props, context) {
-    super(props, context);
-    this.state = {
-      similarRoutes: [],
-      loadingSimilar: true,
-    };
-    if (this.context.config.showSimilarRoutesOnRouteDropDown) {
-      this.fetchSimilarRoutes(this.props.route);
-    }
-  }
-
-  static propTypes = {
-    params: PropTypes.shape({
-      patternId: PropTypes.string.isRequired,
-    }).isRequired,
-    className: PropTypes.string.isRequired,
-    route: routeShape.isRequired,
-    onSelectChange: PropTypes.func.isRequired,
-    serviceDay: PropTypes.string.isRequired,
-    gtfsId: PropTypes.string.isRequired,
-    useCurrentTime: PropTypes.bool,
-    lang: PropTypes.string.isRequired,
-    relayEnvironment: relayShape.isRequired,
-  };
-
-  static defaultProps = {
-    useCurrentTime: false,
-  };
-
-  static contextTypes = {
-    router: routerShape.isRequired,
-    config: configShape,
-    getStore: PropTypes.func.isRequired,
-    intl: intlShape.isRequired,
-  };
-
-  fetchSimilarRoutes = route => {
-    let searchSimilarTo = route.shortName;
-    const c = route.shortName.length ? route.shortName[0] : '';
-    if (c < '0' || c > '9') {
-      // must start with number
-      return;
-    }
-    if (Number.isNaN(Number(route.shortName))) {
-      searchSimilarTo = route.shortName.replace(/\D/g, ''); // Delete all non-digits from the string
-    }
-    if (!searchSimilarTo) {
-      // Dont try to search similar routes for routes that are named with letters (eg. P train)
-      return;
-    }
-    const query = graphql`
-      query RoutePatternSelect_similarRoutesQuery($name: String) {
-        routes(name: $name) {
-          gtfsId
-          shortName
-          longName
-          mode
-          color
-        }
-      }
-    `;
-
-    const params = { name: searchSimilarTo };
-    fetchQuery(this.props.relayEnvironment, query, params, {
-      force: true,
-    })
-      .toPromise()
-      .then(results => {
-        this.setState({
-          similarRoutes: filterSimilarRoutes(results.routes, this.props.route),
-          loadingSimilar: false,
-        });
-      });
-  };
-
-  getOptions = () => {
-    const { gtfsId, params, route } = this.props;
-    const { router } = this.context;
-    const { patterns } = route;
-
-    if (patterns.length === 0) {
-      return null;
-    }
-
-    const futureTrips = enrichPatterns(
-      patterns,
-      false,
-      this.context.config.itinerary.serviceTimeRange,
-    );
-
-    if (futureTrips.length === 0) {
-      return null;
-    }
-
-    const options = sortBy(
-      sortBy(futureTrips, 'inFuture').reverse(),
-      'countTripsForDate',
-    ).reverse();
-    if (options.every(o => o.code !== params.patternId)) {
-      router.replace(routePagePath(gtfsId, PREFIX_STOPS, options[0].code));
-    }
-    return options;
-  };
-
-  render() {
-    const { intl } = this.context;
-    const options = this.getOptions();
-    const currentPattern = options.find(
-      o => o.code === this.props.params.patternId,
-    );
-
-    const possibleMainRoutes = options.slice(0, 2).filter(o => !o.inFuture);
-    let mainRoutes = options.slice(0, 2).filter(o => !o.inFuture);
-    if (
-      possibleMainRoutes.every(o => o.directionId === 0) ||
-      possibleMainRoutes.every(o => o.directionId === 1)
-    ) {
-      mainRoutes = possibleMainRoutes.slice(0, 1);
-    }
-    const specialRoutes = options
-      .slice(mainRoutes.length)
-      .filter(o => !o.inFuture);
-    const futureRoutes = options
-      .slice(mainRoutes.length)
-      .filter(o => o.inFuture);
-
-    const noSpecialRoutes = !specialRoutes || specialRoutes.length === 0;
-    const noFutureRoutes = !futureRoutes || futureRoutes.length === 0;
-    const noSimilarRoutes =
-      !this.state.similarRoutes || this.state.similarRoutes.length === 0;
-
-    const renderButtonOnly =
-      mainRoutes &&
-      mainRoutes.length > 0 &&
-      mainRoutes.length <= 2 &&
-      noSpecialRoutes &&
-      noFutureRoutes &&
-      noSimilarRoutes;
-
-    const directionSwap = mainRoutes.length === 2;
-    if (renderButtonOnly) {
-      const otherPattern = mainRoutes.find(
-        o => o.code !== this.props.params.patternId,
-      );
-      return (
-        <div
-          className={cx('route-pattern-select', this.props.className)}
-          aria-atomic="true"
-        >
-          <label htmlFor="route-pattern-toggle-button">
-            {directionSwap && (
-              <span className="sr-only">
-                <FormattedMessage id="swap-order-button-label" />
-              </span>
-            )}
-            <button
-              id="route-pattern-toggle-button"
-              className="route-pattern-toggle"
-              type="button"
-              onClick={() =>
-                directionSwap
-                  ? this.props.onSelectChange(otherPattern.code)
-                  : null
-              }
-            >
-              {patternTextWithIcon(currentPattern)}
-              {directionSwap && (
-                <Icon className="toggle-icon" img="icon_direction-c" />
-              )}
-            </button>
-          </label>
-        </div>
-      );
-    }
-
-    const optionArray = [];
-    if (mainRoutes.length > 0) {
-      optionArray.push({ options: mainRoutes, name: '' });
-    }
-    if (specialRoutes.length > 0) {
-      optionArray.push({
-        options: specialRoutes,
-        name: intl.formatMessage({
-          id: 'route-page.special-routes',
-        }),
-      });
-    }
-    if (futureRoutes.length > 0) {
-      optionArray.push({
-        options: futureRoutes,
-        name: intl.formatMessage({
-          id: 'route-page.future-routes',
-        }),
-      });
-    }
-
-    if (
-      this.context.config.showSimilarRoutesOnRouteDropDown &&
-      !this.state.loadingSimilar &&
-      this.state.similarRoutes?.length > 0
-    ) {
-      optionArray.push({
-        options: this.state.similarRoutes,
-        name: intl.formatMessage({
-          id: 'route-page.similar-routes',
-        }),
-      });
-    }
-
-    const sectionTitleFontWeight =
-      this.context.config.appBarStyle === 'hsl' ? 500 : 600;
-
-    return (
-      <div
-        className={cx('route-pattern-select', this.props.className)}
-        aria-atomic="true"
-        style={{
-          '--sectionTitleFontWeight': `${sectionTitleFontWeight}`,
-        }}
-      >
-        {/* eslint-disable-next-line  jsx-a11y/label-has-associated-control */}
-        <label htmlFor="select-route-pattern-input">
-          <span className="sr-only">
-            <FormattedMessage id="route-page.pattern-select-title" />
-          </span>
-          <Autosuggest
-            id="select-route-pattern"
-            suggestions={optionArray}
-            multiSection
-            renderSectionTitle={s => {
-              return s.name || null;
-            }}
-            getSectionSuggestions={s => {
-              return s.options;
-            }}
-            getSuggestionValue={s => s.code}
-            renderSuggestion={s =>
-              renderPatternSelectSuggestion(s, currentPattern)
-            }
-            onSuggestionsFetchRequested={() => null}
-            shouldRenderSuggestions={() => true}
-            onSuggestionSelected={(e, { suggestion, suggestionValue }) => {
-              if (!suggestionValue && suggestion.gtfsId) {
-                // for similarRoute links to work when selected with keyboard
-                this.context.router.push(routePagePath(suggestion.gtfsId));
-              }
-            }}
-            inputProps={{
-              value: this.props.params && patternOptionText(currentPattern),
-              onChange: (_, { newValue, method }) => {
-                if (['click', 'enter'].includes(method)) {
-                  this.props.onSelectChange(newValue);
-                }
-              },
-              onMouseDown: () => {
-                addAnalyticsEvent({
-                  category: 'Route',
-                  action: 'OpenDirectionMenu',
-                  name: null,
-                });
-              },
-              id: 'select-route-pattern-input',
-              'aria-autocomplete': 'none',
-              readOnly: true,
-            }}
-            renderInputComponent={inputProps => {
-              return (
-                <>
-                  <div className="input-display" aria-hidden="true">
-                    {patternTextWithIcon(currentPattern)}
-                    <Icon
-                      className="dropdown-arrow"
-                      img="icon_arrow-collapse"
-                    />
-                  </div>
-                  <input {...inputProps} />
-                </>
-              );
-            }}
-            focusInputOnSuggestionClick={false}
-            onSuggestionsClearRequested={() => null}
-          />
-        </label>
-      </div>
-    );
-  }
-}
-
-const storeComponent = connectToStores(
-  props => (
-    <ReactRelayContext.Consumer>
-      {({ environment }) => (
-        <RoutePatternSelect {...props} relayEnvironment={environment} />
-      )}
-    </ReactRelayContext.Consumer>
-  ),
-  ['PreferencesStore'],
-  context => ({
-    serviceDay: unixToYYYYMMDD(
-      context.getStore('TimeStore').getCurrentTime(),
-      context.config,
-    ),
-    lang: context.getStore('PreferencesStore').getLanguage(),
-  }),
-);
-
-storeComponent.contextTypes = {
-  getStore: PropTypes.func.isRequired,
-  config: configShape.isRequired,
 };
 
-const withStore = createRefetchContainer(
-  storeComponent,
-  {
-    route: graphql`
-      fragment RoutePatternSelect_route on Route
-      @argumentDefinitions(date: { type: "String" }) {
-        shortName
-        mode
-        gtfsId
-        patterns {
-          code
-          directionId
-          headsign
-          stops {
-            name
-          }
-          activeDates: trips {
-            serviceId
-            day: activeDates
-          }
-          tripsForDate: tripsForDate(serviceDate: $date) {
-            stoptimes: stoptimesForDate(serviceDate: $date) {
-              scheduledDeparture
-              serviceDay
-            }
-          }
-        }
-      }
-    `,
-  },
-  graphql`
-    query RoutePatternSelectQuery($routeId: String!, $date: String!) {
-      route(id: $routeId) {
-        ...RoutePatternSelect_route @arguments(date: $date)
-      }
-    }
-  `,
-);
+PatternOption.propTypes = {
+  option: patternShape.isRequired,
+  optionIndexTable: PropTypes.shape.isRequired,
+  highlightedIndex: PropTypes.number.isRequired,
+  getItemProps: PropTypes.func.isRequired,
+  currentPattern: patternShape.isRequired,
+};
 
-export { withStore as default, RoutePatternSelect as Component };
+export default function RoutePatternSelect(
+  { currentPattern, optionArray, onSelectChange, className },
+  { config },
+) {
+  // flatten optionArray to an ungrouped 1-D array
+  const flattenedOptions = optionArray.reduce(
+    (options, group) => options.concat(group.options),
+    [],
+  );
+  // lookup table for getting the index of an option, used for highlighting
+  const optionIndexTable = flattenedOptions.reduce(
+    (map, { code, gtfsId }, index) => {
+      // eslint-disable-next-line no-param-reassign
+      map[code || gtfsId] = index;
+      return map;
+    },
+    {},
+  );
+
+  const {
+    isOpen,
+    highlightedIndex,
+    getToggleButtonProps,
+    getMenuProps,
+    getItemProps,
+    getLabelProps,
+  } = useSelect({
+    selectedItem: currentPattern,
+    items: flattenedOptions,
+    onSelectedItemChange: ({ selectedItem }) =>
+      onSelectChange(selectedItem.code),
+    onIsOpenChange: () =>
+      addAnalyticsEvent({
+        category: 'Route',
+        action: 'OpenDirectionMenu',
+        name: null,
+      }),
+    labelId: 'route-pattern-select-label',
+    menuId: 'route-pattern-select-menu',
+    toggleButtonId: 'route-pattern-select-toggle',
+  });
+
+  const sectionTitleFontWeight = config.appBarStyle === 'hsl' ? 500 : 600;
+  return (
+    <div
+      className={cx('route-pattern-select', className)}
+      aria-atomic="true"
+      style={{
+        '--sectionTitleFontWeight': `${sectionTitleFontWeight}`,
+      }}
+    >
+      <div
+        className={cx(
+          'pattern-select-container',
+          isOpen && 'pattern-select-container--open',
+        )}
+      >
+        <div>
+          <label {...getLabelProps()}>
+            <span tabIndex={-1} className="sr-only">
+              {patternOptionText(currentPattern)}
+            </span>
+            <span className="sr-only">
+              <FormattedMessage id="route-page.pattern-select-title" />
+            </span>
+          </label>
+          <div {...getToggleButtonProps()}>
+            <span>
+              <div className="input-display" aria-hidden="true">
+                {patternTextWithIcon(currentPattern)}
+                <Icon className="dropdown-arrow" img="icon_arrow-collapse" />
+              </div>
+            </span>
+          </div>
+        </div>
+
+        <div
+          className={cx(
+            'suggestions-container',
+            isOpen && 'suggestions-container--open',
+          )}
+          hidden={!isOpen}
+          {...getMenuProps({})}
+        >
+          {optionArray.map(section => {
+            return (
+              <div key={section.name} className="section-container">
+                {section.name && (
+                  <div className="section-title">{section.name}</div>
+                )}
+                <ul aria-hidden="true">
+                  {section.options.map(option => (
+                    <PatternOption
+                      key={option.code}
+                      option={option}
+                      optionIndexTable={optionIndexTable}
+                      highlightedIndex={highlightedIndex}
+                      getItemProps={getItemProps}
+                      currentPattern={currentPattern}
+                    />
+                  ))}
+                </ul>
+              </div>
+            );
+          })}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+RoutePatternSelect.propTypes = {
+  currentPattern: patternShape.isRequired,
+  optionArray: PropTypes.arrayOf(
+    PropTypes.shape({
+      name: PropTypes.string,
+      option: patternShape.isRequired,
+    }),
+  ).isRequired,
+  onSelectChange: PropTypes.func.isRequired,
+  className: PropTypes.string.isRequired,
+};
+
+RoutePatternSelect.contextTypes = {
+  config: configShape.isRequired,
+};
