+++ app/component/routepage/route.scss
@@ -1255,7 +1255,7 @@ $margin-left-right: 21.3px;
     width: 24px;
   }
 
-  .react-autosuggest__container {
+  .pattern-select-container {
     width: 100%;
     min-height: 44px;
     background: white;
@@ -1280,7 +1280,7 @@ $margin-left-right: 21.3px;
       padding: 13px 32px 13px 16px;
     }
 
-    .react-autosuggest__input {
+    .select-route-pattern-input {
       height: 0;
       margin: 0;
       padding: 0;
@@ -1292,17 +1292,17 @@ $margin-left-right: 21.3px;
       }
     }
 
-    &.react-autosuggest__container--open .dropdown-arrow {
+    &.pattern-select-container--open .dropdown-arrow {
       transform: rotate(180deg);
     }
 
-    .react-autosuggest__suggestions-container--open {
+    .suggestions-container--open {
       border: 1px solid #ddd;
       border-bottom: none;
       box-shadow: 0 4px 13px rgba(0, 0, 0, 0.15);
     }
 
-    .react-autosuggest__suggestions-container {
+    .suggestions-container {
       background: $background-color-lighter;
       z-index: index($zindex, 'autosuggest-suggestion-container');
       position: absolute;
@@ -1311,7 +1311,7 @@ $margin-left-right: 21.3px;
       right: 0;
       border-radius: 4px;
 
-      .react-autosuggest__section-container {
+      .section-container {
         &:not(:first-child) {
           margin-top: 4px;
           border-top: 1px solid #ddd;
@@ -1331,7 +1331,7 @@ $margin-left-right: 21.3px;
         line-height: 18px;
         letter-spacing: -0.03em;
 
-        .react-autosuggest__section-title {
+        .section-title {
           font-weight: var(--sectionTitleFontWeight);
           border-bottom: 1px solid #ddd;
           min-height: 44px;
@@ -1342,7 +1342,7 @@ $margin-left-right: 21.3px;
           padding: 0;
           margin: 0;
 
-          .react-autosuggest__suggestion {
+          .suggestion {
             border-bottom: 1px solid #ddd;
             list-style: none;
             min-height: 44px;
@@ -1359,7 +1359,7 @@ $margin-left-right: 21.3px;
             }
           }
 
-          .react-autosuggest__suggestion--highlighted {
+          .suggestion--highlighted {
             background: #dfe5e8;
           }
         }
