+++ package.json
@@ -150,6 +150,7 @@
     "connect-redis": "5.0.0",
     "cookie-parser": "1.4.7",
     "debug": "4.3.4",
+    "downshift": "^9.0.10",
     "express": "4.21.2",
     "express-http-proxy": "1.6.3",
     "express-session": "1.18.1",
